Incorrect offset mask for k-mers larger than 15
The determination of the 64-bit mask used in seq_to_hash() was
incorrect due to being based on shifting 1 left by 2k bits.

(1 << (2*k))

This is incorrect because '1' will default to a 32 bit value with
the resulting value being undefined for shifts larger than 31.

This has been fixed by shifting an explicitly defined 64 bit value.